change float for double for more precision

+++ src/main/Configuration.java
@@ -2,6 +2,7 @@ package main;
 
 import main.util.AudioConfig;
 import main.util.SpriteConfig;
+import org.joml.Vector2d;
 import org.joml.Vector2f;
 import org.joml.Vector3f;
 import org.joml.Vector4f;
@@ -68,7 +69,7 @@ public class Configuration {
 
     public static Vector4f hoverGizmoColor = new Vector4f(0f, 0.5f, 0f, 1f);
     public static Vector4f gizmoColor = new Vector4f(0.0f, .0f, 0.0f, 1f);
-    public static Vector2f gizmoScale = new Vector2f(30, 60);
+    public static Vector2d gizmoScale = new Vector2d(30, 60);
     public static int xGizmoXAxis = 10;
     public static int xGizmoYAxis = -9;
     public static int xGizmoRotation = 90;
